test: reduce nondeterminism in the tx expiration test (#1159)
The fixtures in
packages/core-transaction-pool-mem/__tests__/__fixtures__/ contain
transactions that are set to expire after 5 seconds:
  expiration: slots.getTime() + 5
but this is evaluated as soon as jest starts and it takes some time to
set up the test environment which may be more than 5 seconds. If this
happens then the transactions are already expired by the time the
expiration test starts.

To fix this set the expiration from within the test, just before
executing it.