refactor(mempool): remove forged transactions at startup
Checking which transactions from a given set are forged involves
database access (PostgreSQL). We used to remove all forged transactions
every time we derive a list of transactions to forge. This is not
necessary as during normal workflow a transaction is removed from the
pool after being forged.

A scenario where a forged transaction may be in the mem pool is
if it gets forged, then gets deleted from the pool, but the pool does
not immediately sync to disk and then the node crashes. When it starts
the next time and loads the pool contents from disk, then the forged
and removed transaction will resurface. Thus we remove all forged
transactions during startup.