test(mempool): fix test "amount > wallet balance" which was testing nothing
The test was adding an invalid transaction to the pool (invalid because
of not enough funds) and was expecting that transaction not to be
returned by getTransactionsForForging(). The problem is that the method
was called with an argument of 0: getTransactionsForForging(blockSize=0).
The blockSize argument denotes the maximum number of transactions to
return and using 0 would always cause an empty array to be returned.

To fix this use some arbitrary number, larger than 0. Also, extend the
test in two ways:
* Add one valid transaction from another sender to verify that the
  removal of the invalid one will not also ditch the valid one.
* Add one valid transaction from the same sender - it should also be
  removed because all transactions from the same sender are removed upon
  encountering an invalid transaction.