refacor: avoid the non-deterministic setTimeout
Using setTimeout() to "wait" for the log to be flushed is
non-deterministic. It may happen that process.exit() is called before
the log is flushed. Also it has another problem that other error
messages are printed because the process continues execution.

Instead log errors and warnings to stderr instead of stdout. For some
reason stderr is properly flushed before the process is terminated by
process.exit() (whereas stdout is not). Also, error messages belong to
stderr anyway.