perf: optimize the way we estimate the number of rows
When we use OFFSET and LIMIT in SQL we want to assess the total number
of rows that would have been returned if OFFSET and LIMIT were omitted.

Instead of doing a separate query with COUNT(*), parse the output
of EXPLAIN SELECT ... and fetch the number of rows from there.