fix(core-p2p): malformed condition for filtering peers
Currently we would leave in the list all peers that:
"are not me or have invalid port or have invalid version"
the intention must have been:
"are not me and have valid port and have valid version".

This change was first done in 0c2319649 and was later reverted in
109a4d311 with a scarce explanation: "it randomly causes issues". It is
not clear what those issues were.

However the current code is strikingly and obviously wrong - it would
add peers that have invalid port, for example.

So, fix the wrong code and followup by fixing any issues that arise, but
not by resurrecting the wrong code.