fix(core-database-postgres): off-by-one error in getBlocks()
The getBlocks() function takes offset,limit arguments, presumably with
the intention of returning maximum `limit` number of elements, starting
from `offset`. However it would return `limit + 1` elements. E.g. if
given offset=5,limit=2 it would return three elements: 5,6,7 instead of
the expected two 5,6.

As a side effect of this, the `if (blocks.length !== limit)` check
inside getBlocks() would never be true and we would always retrieve the
blocks from the database, even if they are present in the "state".

The callers of getBlocks() were adapted to this wretched behavior and
did ask for `limit - 1`, knowing they would get one more. Adjust that,
now that it is not needed.

Also, rename a confusingly named variable `blocksToRemove` which
actually does not contain the blocks to be removed (!).