feat(core-p2p): Obey ping timeout during peer verification
With this patch the entire peer verification (which is done during
ping()) must complete within the ping timeout. Maybe this is too
aggressive, leaving too little time for verification because the ping
timeout was intended as a timeout for a single network call to
`/peer/status`. However is a good starting point, which may be bumped if
it shows to be too low during testing.