fix(core-api): proper hex validation (#2093)
Buffer.from(value, "hex") will only throw an exception if value is not a
string, according to
https://nodejs.org/api/buffer.html#buffer_class_method_buffer_from_string_encoding
Indeed Buffer.from('foo', 'hex') does not throw. The schema has defined
`type: "string"`, which means that, in all cases when the validate
method is called, it will return true.

Use a proper validation using regular expression and a length check.