feat(core-p2p): Fetch list of peers from at least a few others
Previously we would fetch the list of peers from a single other peer and
be satisfied with it as long as our list of peers satisfies
hasMinimumPeers().

With this change we will query at least 4 peers and combine their
lists. This will help in getting a better view of the network and all of
its peers in times of splits and bad network connectivity.