refactor(core-forger): Increase timeout for getting the network state
The 2 seconds timeout is too short and can unnecessary blow up the
forging process in cases where it is possible to forge successfully.

Set it to an arbitrary 7 seconds which is block time minus 1 second for
other activities.

Elaborate log messages by explicitly saying "Will not forge" + an
explanation "why" when the code would decide not to forge.