fix(core): parse only the last line of pm2 stdout
Fixes instances of "Error: The "XXX" process has entered an unknown state. (undefined)"

This is because sometimes the in-memory pm2 gets out of date (e.g. if multiple pm2 instances are installed on the system). When this happens, an error is printed to stdout prior to the JSON from pm2 jlist being dumped. This prevents JSON.parse from parsing the response properly. By only processing the last line of the output from pm2 jlist, we exclude the error text, so JSON.parse succeeds.