refactor(core-transaction-pool): simplify by not using insertion order
With static fees we wanted to process transactions in insertion order -
"first in, first out" because all of them had the same fee. But with
dynamic fees we do sort them by fee and process the ones with the
highest fee first. We only sorted by insertion order the transactions
with the same fee, but this patch replaces it with sorting by
expiration and completely abolishes the insertion order as it adds a
layer of complexity to the entire transaction pool code.