feat(core-transaction-pool): sort transactions by fee, nonce
Sort transactions in the pool by fee (highest fee first), as before.

However, if we find a transaction from the same sender with a higher
nonce earlier in the sorted array, then move that transaction later in
the array. Ie don't allow transactions with higher nonce before
transactions with a lower nonce (for the same sender).