refactor(core-transaction-pool): expire transactions that don't have … (#2672)
* refactor(core-transaction-pool): expire transactions that don't have expiration

If a transaction does not have an expiration value set by the user, then
we would keep it for a long time in the pool and also it could possibly
be re-sent by somebody (not the author) long time after it was created,
if it has not been forged yet.

Change this to expire after 6h even transactions that don't have an
user-set expiration. 6h considered since the transaction creation time.

* Ignore the expiration of v1 transactions

* Rename some variables and a method

* Properly call the Memory constructor from tests

* test(core-transaction-pool): adjust expiration test

* chore(core-transaction-pool): silence tslint

* test(core-blockchain): disable transactions expiration

Since we try to apply the genesis transactions, which are created a few
years ago, set the pool default expiration to some point in the distant
future.