fix(core-snapshots): do not restore genesis block with wrong id
For some reason the hardcoded id of the genesis block in
packages/crypto/src/networks/*/genesisBlock.json is different than the
one that would be calculated by the crypto package (block
serialize/deserialize).

Because the id as in genesisBlock.json is already in previous_block of
the subsequent block (at height 2), we must use the hardcoded id instead
of the one derived by the crypto package.