fix: try harder to return the requested number of transactions
Before this patch we would take the top "size" transactions, filter away
some of them and return the result, possibly returning less than the
requested "size" even if there could be legit transactions in the pool
that could be returned, up to "size".

Change the code to keep iterating until we have collected "size" valid
transactions, or the end of the pool is reached.