feat(core-api): make it configurable whether to use estimates
Introduce a new boolean config option totalCountIsEstimate and use
database estimates for the total number of rows if it is true (fast)
or use the precise COUNT(*) if the option is false (slow).

So, it is up to the node operator to configure their node for accuracy
vs speed. Add a new property in the response to indicate which one is
being used.

Resolves https://github.com/ArkEcosystem/core/issues/2676
Pagination error in api/blocks/{id}/transactions endpoint