fix(core-blockchain): remove too early check for a chained block
In processBlocks() we would check whether the first downloaded block is
chained and if not then we would return from processBlocks(). However,
if that happens it leads to an "infinite" loop of downloading blocks,
not applying any of them until the nodejs process runs out of memory to
store all the blocks it has downloaded. A block that would cause this is
(on devnet): height=1199883, id=5174570296595098048.

Remove this early check. Later all blocks (not only the first one) are
verified whether chained or not.