misc(core-blockchain): log the reason for discarding a block (#2903)
* Print the reason for discarding a block

When processing blocks, if we find that some block is not chained then
we abandon the rest of the blocks. In this case, print a detailed
message explaining that a block is not chained and why.

Without this printout sync failures could be very obscure and difficult
to analyze.

* Reset last downloaded block if we get an unchained block

This would make us try to download the block at the same height again
from a different peer.

Also, check if the block being tested for isBlockChained() is in the
list of exceptions and if so, assume the check has passed.