perf: Avoid O(m*n) when filtering pool txs and simplify the code
validateTransactions() would take an array of transaction objects and
return a subset of it, but just the serialized property, not the entire
transaction object.

The two callers of validateTransactions() would then filter the array
they provided to validateTransactions() so that it contains only the
returned valid transactions. But, given that the returned array is a
subset of the provided array that means that the returned array is
the same as the filtered input. On top of this the unnecessary filtering
was done using an ineffective O(m*n) algorithm.

So, change validateTransactions() to return the transaction objects
instead of just the serialized property and use the returned array
directly (avoid the unnecessary filtering altogether).