refactor(pool): remove unnecessary check from validateTransactions()
We would call

  handler.throwIfCannotBeApplied(transaction, sender, databaseWalletManager);

followed by

  handler.applyToSender(transaction, localWalletManager);

but applyToSender() calls throwIfCannotBeApplied() internally, so we
don't need to call it ourselves. Also, it was wrong to use the
databaseWalletManager in that call because we are applying the
transactions to localWalletManager, so the databaseWalletManager will be
outdated (e.g. sender nonce and balance will be stale in
databaseWalletManager).

So, remove the unnecessary call to throwIfCannotBeApplied().