fix(core-snapshots): avoid storing duplicate rounds
When appending to a snapshot, in 50 of 51 cases it will happen that the
last round from the appended-to snapshot is the same as the first round
in the being-appended snapshot. Dont store that twice in the snapshot
because then it cannot be restored due to duplicate key error from the
database.