fix(core-snapshots): remove bogus skipRoundRows (#2973)
* fix(core-snapshots): remove bogus skipRoundRows

skipRoundRows was only set to something else than 0 when appending to an
existent snapshot and then it was set to the number of rows in the
rounds table in the appended-to snapshot. Then it would translate to an
OFFSET in the SQL query:

SELECT round, balance, public_key
FROM rounds
WHERE round BETWEEN 9804 AND 9902
ORDER BY round, balance DESC, public_key
OFFSET 500004

but the proper rows to be selected are already defined by the BETWEEN
clause.

So, the usage of OFFSET is wrong and leads to bogus results - when
appending to a snapshot the result would be a snapshot with missing rows
from the rounds table, and subsequently when that snapshot is restored -
gaps (missing rows) in the rounds table in the database.

* fix(core-snapshots): avoid storing duplicate rounds

When appending to a snapshot, in 50 of 51 cases it will happen that the
last round from the appended-to snapshot is the same as the first round
in the being-appended snapshot. Dont store that twice in the snapshot
because then it cannot be restored due to duplicate key error from the
database.