refactor(core): use default heap size regardless of available memory
This was added to trigger the GC sooner on nodes with low memory,
as they tend to run out of memory during a long running sync.
However, all this did was to just work around a symptom and not the
actual issue (*cough* Socketcluster timeouts something*cough*). Now it
is causing issues during bootstrap, because V8 hits the allocation limit
and just kills itself. Since the former is the latter evil, rather take
the risk to run out of memory during a long running sync than during
bootstrap.