perf(core-transactions): disable TransactionReader
Disable the TransactionReader because it inflicts a minor performance
regression. However, do not completely delete it from the source code
because we may revisit it later when the database queries start
returning way too many rows to the nodejs app at once. The proper
solution is to execute the database query once and serve the results to
the app on portions using cursors, rather than executing the same query
multiple times and chopping the portions using OFFSET,LIMIT.