fix: mainnet address exceptions
A handful of mainnet transactions have an invalid recipient. Due to a
refactor of the Address network byte validation it is no longer
trivially possible to handle them. If an invalid address is encountered
during transfer serialization, the error is bubbled up to defer the
`AddressNetworkByteError` until the actual id is available to call
`isException`.