chore(core-database-postgres): convert BYTEA to/from string
Engage pg-promise's ability to modify the values retrieved from the
database on the fly (the cb argument of
https://vitaly-t.github.io/pg-promise/Database.html#result).

This way we convert transparently the vendorField column which comes as
a Buffer from the database to an UTF8 string. In the other direction,
when inserting into the database, the string is converted to a Buffer
using the init property of the column definition.

Rename findMany and findManyWithCount to findAny and findAnyWithCount.
pg-promise has a method many() which throws if no rows are returned. It
also has a method any() which accepts 0, 1 or more rows as a result
(same as manyOrNone() or result()). Our findMany* methods do
not call the many() method but rather accept any number of rows as a
result.