chore(core-database-postgres): convert BYTEA to/from string (#3058)
* chore(core-database-postgres): convert BYTEA to/from string

Rename findMany and findManyWithCount to findAny and findAnyWithCount.
pg-promise has a method many() which throws if no rows are returned. It
also has a method any() which accepts 0, 1 or more rows as a result
(same as manyOrNone() or result()). Our findMany* methods do
not call the many() method but rather accept any number of rows as a
result.