fix(core-p2p): raise `getCommonBlocks` rate limit (#3069)
The peer verifier ends up calling `getCommonBlocks` more often than the
rate limit allows. This is because the amount of calls depends on the
input of the prober - which basically runs `log8(someHeight)` times
and it happens to be more than 5 times. Raising the rate limit from 5
to 9 makes it so that the number of calls are even at height 100M within
the rate limit.