fix(core-blockchain): write accepted blocks to database before… (#3155)
* fix: dont rollback while accepted blocks have not been written to database yet

The `AcceptBlockHandler` already alters `blocksInCurrentRound` while blocks
aren't written to database until the current batch finished processing.

In the case of a fork, we'd stop while blocks have been accepted, but not
written to database yet. This led to inconsistencies and resulted
in an assert being triggered which then wrecked the state machine
effectively stopping the node.

```
AssertionError [ERR_ASSERTION]: The expression evaluated to a falsy value:
  assert_1.default(this.blocksInCurrentRound.pop().data.id === block.data.id)
    at DatabaseService.revertBlock (/home/ark/.config/yarn/global/node_modules/@arkecosystem/core-database/dist/database-service.js:312:25)
    at async revertLastBlock (/home/ark/.config/yarn/global/node_modules/@arkecosystem/core-blockchain/dist/blockchain.js:236:13)
    at async __removeBlocks (/home/ark/.config/yarn/global/node_modules/@arkecosystem/core-blockchain/dist/blockchain.js:260:13)
    at async __removeBlocks (/home/ark/.config/yarn/global/node_modules/@arkecosystem/core-blockchain/dist/blockchain.js:261:13)
    at async Blockchain.removeBlocks (/home/ark/.config/yarn/global/node_modules/@arkecosystem/core-blockchain/dist/blockchain.js:270:9)
    at async startForkRecovery (/home/ark/.config/yarn/global/node_modules/@arkecosystem/core-blockchain/dist/state-machine.js:204:9)
```

This PR ensures that a fork only starts after all accepted blocks are
in the database. Tested manually, before I was able to reproduce it every
time, now I can no longer reproduce it.