feat(core-p2p): implement throttling on outgoing p2p communication (#3170)
Use the rate limiter in reverse to stop ourselves from sending too
many requests to a given peer that could possibly trigger their
protection and result in banning us.

We use the default configuration for rate limits. So this will work as
long as the peer is using the same software and has not changed his
config or has made it more liberal. If the peer has made his limits
more restricted, then we could still trigger their protection.

A more robust solution would be to negotiate the rate limits at peer
handshake and then obey different limits for each peer.