chore(core-database): remove misleading order by expirationValue
expirationValue could be either "height of the blockchain" or "seconds
since the genesis block", so sorting by it gives some misleading
impression that the returned results are sorted by expiration when they
are in fact not.

It is not possible to reliably convert from height to seconds, thus
ditch the ordering by expirationValue and order by lockId which will at
least deterministically return the results in the same order every time.

Closes https://github.com/ArkEcosystem/core/issues/3088