refactor: make the canEnterTransactionPool() interface more robust
Make it impossible for an implementation of canEnterTransactionPool() to
return false (cannot enter) and forget to push an error to the list of
errors. Change is so that it returns an error object, with null return
value meaning "no error, can enter".

Related to https://github.com/ArkEcosystem/core/issues/3322