chore(crypto): add publickey verification
## Summary

Current master/develop only checks for valid hex during PublicKey validation, but does not verify the correctness of a publicKey.

This can lead to the unexpected creation of an unaccessable but technically-"valid" Address.

This PR recommends the following:
- add a 'verify' method to 'class PublicKey' using the already-present 'secp256k1' module
- modify 'Address.fromPublicKey' to drop hex-validation in favor of 'PublicKey.verify'
- add relevant unit-tests

## Additional Comments

- No change of import modules needed
- maintains resolution of circular-dep fixes in c84b976

There are other areas thay would benefit from the 'PublicKey.verify' method; however, no additional checks are being added at this time.

It should also be stated that a developer should use this new method whenever:
- the correctness of a publicKey cannot be guaranteed
- the publicKey is NOT already being checked (`.verify(sig, pubKey, msg)`)